requirements_gujarati.txt
# Gujarati Text Extraction Module Dependencies
# Install with: pip install -r requirements_gujarati.txt

# Core OCR Library
pytesseract>=0.3.10

# Image Processing
Pillow>=10.0.0

# Advanced Image Processing (optional but recommended)
opencv-python>=4.8.0

# Additional utilities
numpy>=1.21.0
